# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkFlow API - Marketing link management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Campaign {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  linkCount: number;
  totalClicks: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCampaignBody {
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
}

export interface UpdateCampaignBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  color?: string;
}

export interface Link {
  id: number;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  campaignName?: string | null;
  title: string;
  slug: string;
  destinationUrl: string;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  utmTerm?: string | null;
  /** @nullable */
  utmContent?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  clickCount: number;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLinkBody {
  /** @nullable */
  campaignId?: number | null;
  title: string;
  /** @nullable */
  slug?: string | null;
  destinationUrl: string;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  utmTerm?: string | null;
  /** @nullable */
  utmContent?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface UpdateLinkBody {
  /** @nullable */
  campaignId?: number | null;
  title?: string;
  destinationUrl?: string;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  utmTerm?: string | null;
  /** @nullable */
  utmContent?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  isActive?: boolean;
}

export interface QrCodeResponse {
  svgDataUrl: string;
  shortUrl: string;
}

export interface AnalyticsSummary {
  totalLinks: number;
  totalClicks: number;
  totalCampaigns: number;
  clicksThisMonth: number;
  clicksLastMonth: number;
  /** @nullable */
  topCampaign?: string | null;
}

export interface ClickDataPoint {
  date: string;
  clicks: number;
}

export interface LinkWithClicks {
  id: number;
  title: string;
  slug: string;
  destinationUrl: string;
  clickCount: number;
  /** @nullable */
  campaignName?: string | null;
}

export type GetLinksParams = {
  campaignId?: number;
};

export type GetClicksOverTimeParams = {
  campaignId?: number;
  linkId?: number;
};

export type GetTopLinksParams = {
  limit?: number;
};