# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LinkFlow API - Marketing link management
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: campaigns
    description: Campaign operations
  - name: links
    description: Link operations
  - name: analytics
    description: Analytics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Campaign deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /links:
    get:
      operationId: getLinks
      tags: [links]
      summary: List all links
      parameters:
        - name: campaignId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a marketing link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkBody"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /links/{id}:
    get:
      operationId: getLink
      tags: [links]
      summary: Get a link by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLink
      tags: [links]
      summary: Update a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkBody"
      responses:
        "200":
          description: Link updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Link deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /links/{id}/qr:
    get:
      operationId: getLinkQr
      tags: [links]
      summary: Get QR code for a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get analytics summary
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"

  /analytics/clicks:
    get:
      operationId: getClicksOverTime
      tags: [analytics]
      summary: Get clicks over time (last 30 days)
      parameters:
        - name: campaignId
          in: query
          required: false
          schema:
            type: integer
        - name: linkId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Click data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClickDataPoint"

  /analytics/top-links:
    get:
      operationId: getTopLinks
      tags: [analytics]
      summary: Get top performing links
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Top links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkWithClicks"

  /r/{slug}:
    get:
      operationId: redirectLink
      tags: [links]
      summary: Redirect a short link (track click and redirect)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to destination URL
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Link expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        linkCount:
          type: integer
        totalClicks:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - linkCount
        - totalClicks
        - createdAt
        - updatedAt

    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
      required:
        - name
        - color

    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string

    Link:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        title:
          type: string
        slug:
          type: string
        destinationUrl:
          type: string
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        utmTerm:
          type: ["string", "null"]
        utmContent:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        clickCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - destinationUrl
        - clickCount
        - isActive
        - createdAt
        - updatedAt

    CreateLinkBody:
      type: object
      properties:
        campaignId:
          type: ["integer", "null"]
        title:
          type: string
        slug:
          type: ["string", "null"]
        destinationUrl:
          type: string
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        utmTerm:
          type: ["string", "null"]
        utmContent:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required:
        - title
        - destinationUrl

    UpdateLinkBody:
      type: object
      properties:
        campaignId:
          type: ["integer", "null"]
        title:
          type: string
        destinationUrl:
          type: string
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        utmTerm:
          type: ["string", "null"]
        utmContent:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean

    QrCodeResponse:
      type: object
      properties:
        svgDataUrl:
          type: string
        shortUrl:
          type: string
      required:
        - svgDataUrl
        - shortUrl

    AnalyticsSummary:
      type: object
      properties:
        totalLinks:
          type: integer
        totalClicks:
          type: integer
        totalCampaigns:
          type: integer
        clicksThisMonth:
          type: integer
        clicksLastMonth:
          type: integer
        topCampaign:
          type: ["string", "null"]
      required:
        - totalLinks
        - totalClicks
        - totalCampaigns
        - clicksThisMonth
        - clicksLastMonth

    ClickDataPoint:
      type: object
      properties:
        date:
          type: string
        clicks:
          type: integer
      required:
        - date
        - clicks

    LinkWithClicks:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        destinationUrl:
          type: string
        clickCount:
          type: integer
        campaignName:
          type: ["string", "null"]
      required:
        - id
        - title
        - slug
        - destinationUrl
        - clickCount