# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkFlow API - Marketing link management
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  totalLinks: number;
  totalClicks: number;
  totalCampaigns: number;
  clicksThisMonth: number;
  clicksLastMonth: number;
  /** @nullable */
  topCampaign?: string | null;
}