# lib/api-zod/src/generated/types/clickDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkFlow API - Marketing link management
 * OpenAPI spec version: 0.1.0
 */

export interface ClickDataPoint {
  date: string;
  clicks: number;
}