# lib/api-zod/src/generated/types/createLinkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkFlow API - Marketing link management
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLinkBody {
  /** @nullable */
  campaignId?: number | null;
  title: string;
  /** @nullable */
  slug?: string | null;
  destinationUrl: string;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  /** @nullable */
  utmTerm?: string | null;
  /** @nullable */
  utmContent?: string | null;
  /** @nullable */
  expiresAt?: Date | null;
}